more optimization in calendar drawing: - calculate the locale-specific hour strings for the legend once, rather than on every paint - use Freeze/Thaw on the header while changing the widths of all the columns - round the 255-based colors to integers early, so that folks dealing with rgb2color results are not bogged down with floats
git-svn-id: http://svn.osafoundation.org/chandler/trunk@7659 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/parcels/osaf/framework/blocks/DrawingUtilities.py
@@ -12,7 +12,7 @@ def color2rgb(red, green, blue):
     return red/255.0, green/255.0, blue/255.0
 
 def rgb2color(r, g, b):
-    return r*255, g*255, b*255
+    return int(r*255), int(g*255), int(b*255)
 
 def SetTextColorsAndFont(grid, attr, dc, isSelected):
     """
